test: Make GetNameServers() utility function port-aware
Also make it generally available.
